requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
click>=8.1.0
pytest>=7.4.0
pytest-mock>=3.11.0
python-dateutil>=2.8.0
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0
requests-html>=0.10.0
cloudscraper>=1.2.71
proxy-randomizer>=1.0.0
fake-useragent>=1.4.0
pycurl>=7.45.0
tor>=0.1.0
